add g:flutter_show_log_on_run and remove global settings
This PR adds the option to leave the `__Flutter_Output__` buffer hidden
when you first run `:FlutterRun`. The buffer is still created in the
background, and is available if you want it. The default behaviour is
to open the buffer.

I've also removed some `setlocal`s from the same part of the code, since
they were modifying global config in an undesirable way (often for no
obvious reason). This was happening because some options do not
have a buffer local value. In these cases, setlocal will simply modify
the global instead (see `:help setl`). Details of removed settings and
why I removed them are here.

* `showcmd`, `noruler`, `noshowmode` and `laststatus` have no local
  values, so global configuration was being modified. It seems like a
  bug that the plugin enforces a specific global configuration that has no
  impact on plugin functionality.
* `hidden` is essentially just the default value for `bufhidden`. It has
  no local value and setting the global value is unnecessary since the
  buffer local `bufhidden=hide` is used for this buffer anyway